Added toxic gas and alien obstacles as well
