+++ android/assets/fire.png
