+++ core/src/com/spacejunk/game/obstacles/ToxicGasObstacle.java
@@ -7,10 +7,11 @@ import com.spacejunk.game.levels.Level;
  * Created by vidxyz on 2/9/18.
  */
 
-public class FireObstacle extends Obstacle {
+public class ToxicGasObstacle extends Obstacle {
 
-    public FireObstacle(Level level) {
-        this.obstacleTexture = new Texture("fire.png");
+
+    public ToxicGasObstacle(Level level) {
+        this.obstacleTexture = new Texture("toxic_gas_green.png");
         this.level = level;
     }
